GCC porting fixes.

+++ include/ofp/types.h
@@ -81,7 +81,7 @@ using Milliseconds = std::chrono::milliseconds;
 
 // Literal type `ms` is not supported until C++14. Use _ms for now.
 
-constexpr Milliseconds operator"" _ms(UInt64 msec) {
+constexpr Milliseconds operator"" _ms(unsigned long long int msec) {
   return Milliseconds{msec};
 }
